requirements.txt
langchain-core~=0.3.41
pillow~=11.1.0
langchain~=0.3.20
PyYAML~=6.0.2